Moving docs and tests around
